Restore one-liner.
The suggested one seems like a completely overengineered solution.
Maybe I simply can't understand the benefits yet.

It loses functionality. The reason for `apropos --long` is
because it allows you to get the full short description without it
being truncated. This means that you can search the complete
short description even if it appears truncated in dmenu.
It also means that each entry will occupy the full width of the
prompt so it'll visually truncate automatically based on font.